requirements.txt
scrapy
fabric==1.10
git+https://github.com/django-nonrel/django@nonrel-1.5
git+https://github.com/django-nonrel/djangotoolbox
git+https://github.com/django-nonrel/mongodb-engine
glue==0.9.4
redis==2.10.3
django-grappelli==2.6.1
south==1.0
scrapy-redis
pymongo==2.7.2